Move operators to grafana/grafana
