+++ pkg/cmd/grafana/main.go
@@ -7,9 +7,9 @@ import (
 	"github.com/fatih/color"
 	"github.com/urfave/cli/v2"
 
-	_ "github.com/grafana/grafana/apps/provisioning/pkg/operators"
 	gcli "github.com/grafana/grafana/pkg/cmd/grafana-cli/commands"
 	"github.com/grafana/grafana/pkg/cmd/grafana-server/commands"
+	_ "github.com/grafana/grafana/pkg/operators"
 	"github.com/grafana/grafana/pkg/server"
 	"github.com/grafana/grafana/pkg/services/apiserver/standalone"
 )
