+++ pkg/operators/README.md
