+++ pkg/operators/register.go
